requirements.txt
asgiref==3.4.1
astroid==2.8.6
async-generator==1.10
attrs==21.2.0
backports.entry-points-selectable==1.1.1
brew==0.1.3
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
distlib==0.3.3
Django==3.2.9
django-ckeditor==6.2.0
django-js-asset==1.2.2
filelock==3.4.0
Flask==2.0.2
Flask-MySQL==1.5.2
h11==0.12.0
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.26
mysqlclient==2.1.0
outcome==1.1.0
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.18.1
pycparser==2.20
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
PyMySQL==1.0.2
pyOpenSSL==21.0.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pytz==2021.3
pyunitreport==0.1.4
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.0
urllib3==1.26.7
virtualenv==20.10.0
Werkzeug==2.0.2
wrapt==1.13.3
wsproto==1.0.0
